requirements.txt
Authlib==0.15.4
Flask==2.0.1
Flask_Compress==1.9.0
Flask_Cors==3.0.10
Flask_HTTPAuth==4.2.0
Flask_Mail==0.9.1
Flask_Reuploaded==0.5.0
Flask_SQLAlchemy==2.5.1
flask_uploads==0.2.1
Flask_WTF==0.14.3
itsdangerous==2.0.1
librosa==0.8.0
matplotlib==3.3.4
numpy==1.19.3
opencv_python==4.5.2.52
pandas==1.2.4
Pillow==9.1.1
pydub==0.24.1
scipy==1.6.3
SQLAlchemy==1.4.17
tensorflow==2.6.0
tornado==6.1
Werkzeug==2.0.1
WTForms==2.3.3
XlsxWriter==1.4.3
